Combine to one file and simplify a little
